Remove <Alert/> and update <Notification/>
