--- @stellar/design-system-website/docs/components/notifications/alert.mdx
@@ -1,25 +1,0 @@
----
-slug: /alert
-description: "Display an alert message inline."
----
-
-# Alert
-
-<ComponentDescription componentName="Alert" />
-
-## Example
-
-```tsx live
-<PreviewBlock componentName="Alert">
-  <Alert variant="primary">
-    Lorem ipsum dolor sit amet consectetur adipisicing elit
-  </Alert>
-</PreviewBlock>
-```
-
-## Props
-
-<ComponentProps
-  componentName="Alert"
-  relatedType={["AlertBaseProps", "AlertActionButton", "AlertActionLink"]}
-/>
