--- @stellar/design-system/src/components/Alert/styles.scss
@@ -1,128 +1,0 @@
-@use "../../utils.scss" as *;
-
-.Alert {
-  --Alert-radius: #{pxToRem(8px)};
-  --Alert-icon-color: var(--Alert-icon-color);
-  background-color: var(--Alert-background-color);
-  border: 1px solid var(--Alert-border-color);
-  border-radius: var(--Alert-radius);
-
-  font-size: pxToRem(14px);
-  line-height: pxToRem(20px);
-  color: var(--sds-clr-gray-11);
-  font-weight: var(--sds-fw-regular);
-
-  display: flex;
-  flex-direction: row;
-  gap: pxToRem(12px);
-  padding: pxToRem(12px);
-
-  .Alert__content__iconContainer {
-    align-self: flex-start;
-  }
-
-  .NavButton {
-    top: pxToRem(-6px);
-    right: pxToRem(-6px);
-    position: absolute;
-  }
-
-  .Alert__close {
-    position: relative;
-    width: pxToRem(24px);
-    flex-shrink: 0;
-    flex-grow: 0;
-  }
-
-  .Alert__content {
-    align-items: center;
-  }
-
-  &__content {
-    flex: 1;
-    display: flex;
-    flex-direction: row;
-    gap: pxToRem(12px);
-
-    &__iconContainer {
-      flex-shrink: 0;
-      flex-grow: 0;
-    }
-
-    &__contentContainer {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      gap: pxToRem(12px);
-    }
-
-    &__textContainer {
-      flex: 1;
-      display: flex;
-      flex-direction: column;
-      gap: pxToRem(4px);
-    }
-
-    &__buttons {
-      display: flex;
-      gap: pxToRem(12px);
-
-      .Link {
-        font-weight: var(--sds-fw-semi-bold);
-      }
-    }
-  }
-
-  &__icon {
-    width: pxToRem(16px);
-    height: pxToRem(18px);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    svg {
-      display: block;
-      width: pxToRem(16px);
-      height: pxToRem(16px);
-      stroke: var(--Alert-icon-color);
-    }
-  }
-
-  // Variant
-  &--primary {
-    --Alert-icon-color: var(--sds-clr-lilac-09);
-    --Alert-background-color: var(--sds-clr-lilac-01);
-    --Alert-border-color: var(--sds-clr-lilac-06);
-  }
-
-  &--success {
-    --Alert-icon-color: var(--sds-clr-green-09);
-    --Alert-background-color: var(--sds-clr-green-01);
-    --Alert-border-color: var(--sds-clr-green-06);
-  }
-
-  &--warning {
-    --Alert-icon-color: var(--sds-clr-amber-09);
-    --Alert-background-color: var(--sds-clr-amber-01);
-    --Alert-border-color: var(--sds-clr-amber-06);
-  }
-
-  &--error {
-    --Alert-icon-color: var(--sds-clr-red-09);
-    --Alert-background-color: var(--sds-clr-red-01);
-    --Alert-border-color: var(--sds-clr-red-06);
-  }
-
-  // SDS overrides
-  .Heading--xs {
-    font-size: pxToRem(14px);
-    line-height: pxToRem(20px);
-    color: var(--sds-clr-gray-12);
-    font-weight: var(--sds-fw-medium);
-    letter-spacing: normal;
-  }
-
-  .NavButton {
-    --NavButton-size: #{pxToRem(32px)};
-  }
-}
