+++ @stellar/design-system/src/components/Notification/index.tsx
@@ -1,9 +1,11 @@
 import React from "react";
+import { Button } from "../Button";
 import { Icon } from "../../icons";
+
 import "./styles.scss";
 
 /** */
-export interface NotificationProps {
+export interface NotificationBaseProps {
   /** Variant of the notification */
   variant: "primary" | "secondary" | "success" | "error" | "warning";
   /** Notification title */
@@ -16,12 +18,21 @@ export interface NotificationProps {
   children?: string | React.ReactNode;
 }
 
-interface Props
-  extends NotificationProps, React.HtmlHTMLAttributes<HTMLDivElement> {
-  title: string;
-}
+/** */
+export type NotificationActionButton = {
+  /** Action function */
+  onAction?: () => void;
+  /** Notification action label @defaultValue `Action` */
+  actionLabel?: string;
+  /** Notification action link */
+  actionLink?: undefined;
+  /** Dismiss and close action */
+  onClose?: () => void;
+};
 
-// TODO: Notification: add action buttons
+export type NotificationProps = NotificationActionButton &
+  NotificationBaseProps;
+
 // TODO: Notification: floating with max width
 
 /**
@@ -33,23 +44,70 @@ export const Notification = ({
   icon,
   isFilled,
   children,
-}: Props): React.ReactElement => {
+  onAction,
+  onClose,
+  actionLink,
+  actionLabel = "Action",
+}: NotificationProps): React.ReactElement => {
   const additionalClasses = [
     `Notification--${variant}`,
     ...(isFilled ? [`Notification--filled`] : []),
   ].join(" ");
 
+  const renderActionElement = () => {
+    if (!(onAction || actionLink)) {
+      return null;
+    }
+
+    return (
+      <Button variant={getActionVariant()} size="md">
+        {actionLabel}
+      </Button>
+    );
+  };
+
+  const renderDismissElement = () => {
+    if (!onClose) {
+      return null;
+    }
+
+    return (
+      <Button variant="tertiary" onClick={onClose} size="md">
+        Dismiss
+      </Button>
+    );
+  };
+
+  const getActionVariant = () => {
+    return variant === "error" ? "error" : "primary";
+  };
+
   return (
     <div className={`Notification ${additionalClasses}`}>
-      <div className="Notification__title">
-        <div className="Notification__title__icon">
-          {icon ? icon : <Icon.InfoCircle />}
+      <div className="Notification__content">
+        <div className="Notification__content__iconContainer">
+          <div className="Notification__icon">
+            {icon ? icon : <Icon.InfoCircle />}
+          </div>
         </div>
-        <div className="Notification__title__text">{title}</div>
+        <div className="Notification__content__contentContainer">
+          <div className="Notification__content__text">
+            <div className="Notification__content__title">{title}</div>
+
+            {children ? (
+              <div className="Notification__content__message">{children}</div>
+            ) : null}
+          </div>
+
+          {onClose || onAction || actionLink ? (
+            <div className="Notification__content__buttons">
+              {renderDismissElement()}
+
+              {renderActionElement()}
+            </div>
+          ) : null}
+        </div>
       </div>
-      {children ? (
-        <div className="Notification__message">{children}</div>
-      ) : null}
     </div>
   );
 };
