+++ @stellar/design-system/src/components/Notification/styles.scss
@@ -14,54 +14,83 @@
   border-radius: pxToRem(8px);
   display: flex;
   flex-direction: column;
-  gap: pxToRem(8px);
+  // gap: pxToRem(8px);
   padding: pxToRem(12px);
   width: 100%;
 
-  &__title {
-    display: flex;
-    align-items: flex-start;
-    gap: pxToRem(8px);
+  &__icon {
+    width: pxToRem(16px);
+    height: pxToRem(16px);
+    margin-top: pxToRem(2px);
+    flex-shrink: 0;
 
-    &__icon {
-      width: pxToRem(16px);
-      height: pxToRem(16px);
-      margin-top: pxToRem(2px);
-      flex-shrink: 0;
-
-      svg {
-        width: 100%;
-        height: 100%;
-        stroke: var(--Notification-color-icon);
-      }
-    }
-
-    &__text {
-      color: var(--sds-clr-gray-12);
-      font-size: pxToRem(14px);
-      line-height: pxToRem(20px);
-      font-weight: var(--sds-fw-semi-bold);
+    svg {
+      width: 100%;
+      height: 100%;
+      stroke: var(--Notification-color-icon);
     }
   }
 
-  &__message {
-    color: var(--sds-clr-gray-11);
-    font-size: pxToRem(14px);
-    line-height: pxToRem(20px);
-    font-weight: var(--sds-fw-regular);
+  &__content {
+    flex: 1;
     display: flex;
-    flex-direction: column;
-    gap: pxToRem(4px);
+    flex-direction: row;
+    gap: pxToRem(12px);
 
-    p {
-      margin-bottom: 0 !important;
+    &__iconContainer {
+      flex-shrink: 0;
+      flex-grow: 0;
     }
 
-    ul {
-      font-size: inherit;
-      line-height: inherit;
-      color: inherit;
-      list-style-type: disc;
+    &__contentContainer {
+      flex: 1;
+      display: flex;
+      flex-direction: column;
+      gap: pxToRem(12px);
+    }
+
+    &__text {
+      display: flex;
+      align-items: flex-start;
+      gap: pxToRem(4px);
+      flex-direction: column;
+    }
+
+    &__title {
+      color: var(--sds-clr-gray-12);
+      font-size: pxToRem(14px);
+      line-height: pxToRem(20px);
+      font-weight: var(--sds-fw-medium);
+    }
+
+    &__message {
+      color: var(--sds-clr-gray-11);
+      font-size: pxToRem(14px);
+      line-height: pxToRem(20px);
+      font-weight: var(--sds-fw-regular);
+      display: flex;
+      flex-direction: column;
+      gap: pxToRem(4px);
+
+      p {
+        margin-bottom: 0 !important;
+      }
+
+      ul {
+        font-size: inherit;
+        line-height: inherit;
+        color: inherit;
+        list-style-type: disc;
+      }
+    }
+
+    &__buttons {
+      display: flex;
+      gap: pxToRem(12px);
+    }
+
+    .NavButton {
+      --NavButton-size: #{pxToRem(32px)};
     }
   }
 
@@ -81,8 +110,8 @@
   }
 
   &--success {
-    --Notification-color-background: var(--sds-clr-gray-01);
-    --Notification-color-border: var(--sds-clr-gray-06);
+    --Notification-color-background: var(--sds-clr-green-01);
+    --Notification-color-border: var(--sds-clr-green-06);
     --Notification-color-icon: var(--sds-clr-green-09);
     --Notification-color-action: var(--sds-clr-green-11);
 
@@ -92,8 +121,8 @@
   }
 
   &--error {
-    --Notification-color-background: var(--sds-clr-gray-01);
-    --Notification-color-border: var(--sds-clr-gray-06);
+    --Notification-color-background: var(--sds-clr-red-01);
+    --Notification-color-border: var(--sds-clr-red-06);
     --Notification-color-icon: var(--sds-clr-red-09);
     --Notification-color-action: var(--sds-clr-red-11);
 
@@ -103,8 +132,8 @@
   }
 
   &--warning {
-    --Notification-color-background: var(--sds-clr-gray-01);
-    --Notification-color-border: var(--sds-clr-gray-06);
+    --Notification-color-background: var(--sds-clr-amber-01);
+    --Notification-color-border: var(--sds-clr-red-06);
     --Notification-color-icon: var(--sds-clr-amber-09);
     --Notification-color-action: var(--sds-clr-amber-11);
